fix blockchain type in indexer, add scripts

+++ Assets/SequenceSDK/SequenceIndexer/Scripts/DataTypes/GetChainIDReturn.cs
@@ -1,8 +1,10 @@
+using System.Numerics;
+
 namespace SequenceSharp
 {
     [System.Serializable]
     public class GetChainIDReturn
     {
-        public int chainID;
+        public BigInteger chainID;
     }
 }
